Enhanced Dynamic Conflicts
Now the system operates on permissive and restrictive commands that understands both shifts and days.

Examples:

Beth N can only work on Monday and Friday

Beth N can't work on Saturday or Sunday

Beth N can only work on Monday, Wednesday and Friday in the afternoon

Beth N cannot work on Thursday in the morning

"cannot work" is the restrictive command and if that is not found, the program assumes its a permissive command.